requirements.txt
streamlit
folium
streamlit-folium
sentinelhub
python-dotenv
pillow
numpy
opencv-python
transformers
torch
requests